requirements.txt
streamlit
PyPDF2
langchain
faiss-cpu
sentence-transformers
chromadb
pyyaml
langchain-community
langchain-ollama
